+++ NEWS.md
@@ -1,5 +1,19 @@
 The current version of the DLMtool package is available for download from [CRAN](https://CRAN.R-project.org/package=DLMtool).
 
+## DLMtool Development Version 
+
+ 
+### Minor changes 
+- add option to specify width of the CAL bins using `OM@cpars$binWidth`
+- change MGT calculation from Z to M
+- add option to hide printed results in `TradePlot`
+
+### Fixes
+- fix indexing issue for recruitment deviations
+- fix issue with `cpars$mov` when nareas>2
+
+
+
 ## DLMtool 5.3 
 
 ### New Features